+++ src/main/java/net/openhft/hashing/CompactLatin1CharSequenceAccess.java
@@ -1,4 +1,4 @@
-/*
+/**
  * Copyright 2013-2025 chronicle.software; SPDX-License-Identifier: Apache-2.0
  */
 package net.openhft.hashing;
@@ -13,69 +13,10 @@ import static java.nio.ByteOrder.BIG_ENDIAN;
 import static java.nio.ByteOrder.LITTLE_ENDIAN;
 import static net.openhft.hashing.UnsafeAccess.BYTE_BASE;
 
-/*
- * Compress Latin1 Access
- *
- * Explaination:
- *
- * compressed idx :  0  1  2  3  4  5
- * compressed bytes: 12 34 56 78 9A BC
- *
- * compressed idx :  0     1     2     3     4     5
- * expanded index :  0  1  2  3  4  5  6  7  8  9  A  B
- * expanded LE mem:  12 00 34 00 56 00 78 00 9A 00 BC 00
- * expanded BE mem:  00 12 00 34 00 56 00 78 00 9A 00 BC
- *   align LE byte:  []    --> 0x12
- *   align BE byte:  []    --> 0x00
- * unalign LE byte:     [] --> 0x00
- * unalign BE byte:     [] --> 0x12
- *
- * compressed idx :  0     1     2     3     4     5
- * expanded index :  0  1  2  3  4  5  6  7  8  9  A  B
- * expanded LE mem:  12 00 34 00 56 00 78 00 9A 00 BC 00
- * expanded BE mem:  00 12 00 34 00 56 00 78 00 9A 00 BC
- *   align LE char:  [___]    --> 0x12
- *   align BE char:  [___]    --> 0x12
- * unalign LE char:     [___] --> 0x3400
- * unalign BE char:     [___] --> 0x1200
- *
- * compressed idx :  0     1     2     3     4     5
- * expanded index :  0  1  2  3  4  5  6  7  8  9  A  B
- * expanded LE mem:  12 00 34 00 56 00 78 00 9A 00 BC 00
- * expanded BE mem:  00 12 00 34 00 56 00 78 00 9A 00 BC
- *   align LE int :  [_________]    --> 0x340012
- *   align BE int :  [_________]    --> 0x120034
- * unalign LE int :     [_________] --> 0x56003400
- * unalign BE int :     [_________] --> 0x12003400
- *
- * compressed idx :  0     1     2     3     4     5
- * expanded index :  0  1  2  3  4  5  6  7  8  9  A  B
- * expanded LE mem:  12 00 34 00 56 00 78 00 9A 00 BC 00
- * expanded BE mem:  00 12 00 34 00 56 00 78 00 9A 00 BC
- *   align LE long:  [_____________________]    --> 0x78005600340012
- *   align BE long:  [_____________________]    --> 0x12003400560078
- * unalign LE long:     [_____________________] --> 0x9A00780056003400
- * unalign BE long:     [_____________________] --> 0x1200340056007800
- *
- * Parameters:
- *
- * Parameters must satisfy: 0 <= offset < offset + typeWidth <= input.length*2.
- * When offset + typeWidth >= (input.length + 1)*2, the behavior is undefined, throwing a exception
- * or returning dirty results.
- * When offset + typeWidth == input.length*2 + 1,
- * 1) on BE machine, the result is correct
- * 2) on LE machine, the behavior is undefined, throwing a exception or returning dirty results.
- *
- * compressed idx :  0
- * expanded index :  0  1
- * expanded LE mem:  12 00
- * expanded BE mem:  00 12
- *   align LE char:  [___]    --> 0x12
- *   align BE char:  [___]    --> 0x12
- * unalign LE char:     [_??] --> 0x??00, exception or dirty
- * unalign BE char:     [_00] --> 0x1200, correct
- *
- * Notes: This access is based on the UnsafeAccess, so only works for the native order.
+/**
+ * Provides a {@link CharSequence}-style view over compact Latin-1 byte arrays. Each input byte is
+ * treated as a UTF-16 code unit with a zero high byte. This access relies on {@link UnsafeAccess}
+ * and only supports the platform native byte order.
  */
 @ParametersAreNonnullByDefault
 public class CompactLatin1CharSequenceAccess extends Access<byte[]> {
@@ -95,6 +36,9 @@ public class CompactLatin1CharSequenceAccess extends Access<byte[]> {
 
     private CompactLatin1CharSequenceAccess() {}
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public long getLong(final byte[] input, final long offset) {
         final long byteIdx = (offset + UNSAFE_IDX_ADJUST) >> 1;
@@ -107,6 +51,7 @@ public class CompactLatin1CharSequenceAccess extends Access<byte[]> {
         return expanded;
     }
 
+    /** {@inheritDoc} */
     @Override
     public int getInt(final byte[] input, final long offset) {
         final long byteIdx = (offset + UNSAFE_IDX_ADJUST) >> 1;
@@ -122,13 +67,14 @@ public class CompactLatin1CharSequenceAccess extends Access<byte[]> {
     public long getUnsignedInt(final byte[] input, final long offset) {
         final long byteIdx = (offset + UNSAFE_IDX_ADJUST) >> 1;
         final int compact = UNSAFE.getShort(input, byteIdx) & 0xFFFF;
-        final long expanded = (long)(((compact << 8) | compact) & 0xFF00FF);
+        final long expanded = ((compact << 8) | compact) & 0xFF00FF;
         if (((int)offset & 1) == 1) {
             return expanded << 8;
         }
         return expanded;
     }
 
+    /** {@inheritDoc} */
     @Override
     public int getShort(final byte[] input, final long offset) {
         if (((int)offset & 1) == 0) {
@@ -140,6 +86,7 @@ public class CompactLatin1CharSequenceAccess extends Access<byte[]> {
         }
     }
 
+    /** {@inheritDoc} */
     @Override
     public int getUnsignedShort(final byte[] input, final long offset) {
         if (((int)offset & 1) == 0) {
@@ -151,15 +98,17 @@ public class CompactLatin1CharSequenceAccess extends Access<byte[]> {
         }
     }
 
+    /** {@inheritDoc} */
     @Override
     public int getByte(final byte[] input, final long offset) {
         if (ARRAY_IDX_ADJUST == ((int)offset & 1)) {
             return 0;
         } else {
-            return (int)input[(int)(offset >> 1)];
+            return input[(int)(offset >> 1)];
         }
     }
 
+    /** {@inheritDoc} */
     @Override
     public int getUnsignedByte(final byte[] input, final long offset) {
         if (ARRAY_IDX_ADJUST == ((int)offset & 1)) {
